Move auth components to auth subdirectory
